+++ src/main/java/site/board/boardtraining/domain/board/dto/business/CreateBoardDto.java
@@ -1,4 +1,4 @@
-package site.board.boardtraining.domain.board.dto;
+package site.board.boardtraining.domain.board.dto.business;
 
 import site.board.boardtraining.domain.board.entity.Board;
 import site.board.boardtraining.domain.member.entity.Member;
