+++ src/main/java/site/board/boardtraining/domain/board/dto/business/DeleteBoardDto.java
@@ -1,4 +1,4 @@
-package site.board.boardtraining.domain.board.dto;
+package site.board.boardtraining.domain.board.dto.business;
 
 public record DeleteBoardDto(
         Long boardId,
